implemented tests for #38, forbidden access check on getToolsByRepo

+++ backend/src/main/java/com/dsec/backend/service/repo/RepoServiceImpl.java
@@ -141,7 +141,7 @@ public class RepoServiceImpl implements RepoService {
         {
             for(Integer toolID : repoToolUpdateDTO.getTools())
             {
-                ToolEntity tool = toolService.getToolByID((int)toolID);
+                ToolEntity tool = toolService.getToolByID(toolID);
                 toolRepoRepository.save(new ToolRepo(null, repo.get(), tool));
             }
         }
